Reset the encryption for travis

+++ scripts/travis_after_success.sh
@@ -11,16 +11,23 @@ if [[ ${TRAVIS_BRANCH} == "master" ]] && [[ ${TRAVIS_PULL_REQUEST} == "false" ]]
   openssl version
   # Decrypt the private key.
   openssl aes-256-cbc -k "$travis_key_password" -md sha256 -d -a -in travis_key.enc -out ./travis_key
-  chmod 400 ./travis_key
-  echo "Host github.com" > ~/.ssh/config
-  echo "  IdentityFile $(pwd)/travis_key" >> ~/.ssh/config
 
-  git clone --bare ${mirror_from} forte_bare
-  if cd forte_bare; then
-    git push --mirror ${mirror_to}
+  if [ -s ./travis_key ]
+  then
+	chmod 400 ./travis_key
+  	echo "Host github.com" > ~/.ssh/config
+  	echo "  IdentityFile $(pwd)/travis_key" >> ~/.ssh/config
+
+  	git clone --bare ${mirror_from} forte_bare
+  	if cd forte_bare; then
+    	git push --mirror ${mirror_to}
+  	else
+   		echo "Cannot cd into forte_bare, clone may be unsuccessful."
+  	fi
   else
-    echo "Cannot cd into forte_bare, clone may be unsuccessful."
+	echo "Decrypted key is empty, will not try to push."
   fi
+   
 else
   echo "Not on the master of the main repo, will not do mirror."
   echo "Travis Branch: ""${TRAVIS_BRANCH}"
